Cast size_t

+++ Src-params/ChanMap.cpp
@@ -62,7 +62,7 @@ ChanMapDesc ChanMapDesc::fromWhSpcSepString( const QString &s_in )
 //
 void ChanMap::defaultOrder( QVector<int> &v ) const
 {
-    int n = e.size();
+    int n = (int)e.size();
 
     v.resize( n );
 
@@ -81,7 +81,7 @@ void ChanMap::defaultOrder( QVector<int> &v ) const
 void ChanMap::userOrder( QVector<int> &v ) const
 {
     QMap<int,int>   order2entry;
-    int             n = e.size(),
+    int             n = (int)e.size(),
                     k = 0;
 
     for( int i = 0; i < n; ++i )
@@ -222,7 +222,7 @@ QString ChanMapNI::toString() const
 {
     QString     s;
     QTextStream ts( &s, QIODevice::WriteOnly );
-    int         n = e.size();
+    int         n = (int)e.size();
 
     ts << "(" << MN << "," << MA << ","
               << C  << "," << XA << "," << XD << ")";
@@ -261,7 +261,7 @@ QString ChanMapNI::toWhSpcSepString() const
 {
     QString     s;
     QTextStream ts( &s, QIODevice::WriteOnly );
-    int         n = e.size();
+    int         n = (int)e.size();
 
     ts << MN << "," << MA << ","
        << C  << "," << XA << "," << XD << "\n";
@@ -378,7 +378,7 @@ QString ChanMapOB::toString() const
 {
     QString     s;
     QTextStream ts( &s, QIODevice::WriteOnly );
-    int         n = e.size();
+    int         n = (int)e.size();
 
     ts << "(" << XA << "," << XD << "," << SY << ")";
 
@@ -415,7 +415,7 @@ QString ChanMapOB::toWhSpcSepString() const
 {
     QString     s;
     QTextStream ts( &s, QIODevice::WriteOnly );
-    int         n = e.size();
+    int         n = (int)e.size();
 
     ts << XA << "," << XD << "," << SY << "\n";
 
@@ -498,7 +498,7 @@ void ChanMapIM::setImroOrder( const IMROTbl *R )
     QVector<int>    v;
     R->eaChansOrder( v );
 
-    for( int i = 0, n = e.size(); i < n; ++i )
+    for( int i = 0, n = (int)e.size(); i < n; ++i )
         e[i].order = v[i];
 }
 
@@ -543,7 +543,7 @@ QString ChanMapIM::toString() const
 {
     QString     s;
     QTextStream ts( &s, QIODevice::WriteOnly );
-    int         n = e.size();
+    int         n = (int)e.size();
 
     ts << "(" << AP << "," << LF << "," << SY << ")";
 
@@ -580,7 +580,7 @@ QString ChanMapIM::toWhSpcSepString() const
 {
     QString     s;
     QTextStream ts( &s, QIODevice::WriteOnly );
-    int         n = e.size();
+    int         n = (int)e.size();
 
     ts << AP << "," << LF << "," << SY << "\n";
